requirements.txt
amqp==2.3.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4
Babel==2.6.0
billiard==3.5.0.4
celery==4.1.1
Click==7.0
Django==2.1.2
django-health-check==3.0.0
django-redis==4.8.0
django-rq==1.2.0
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.9.2
flake8==3.5.0
flower==0.9.2
isort==4.3.4
kombu==4.2.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
pep8==1.7.1
pluggy==0.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.7.0
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==2.1.1
pytest==3.8.2
python-dateutil==2.7.3
pytz==2018.5
redis==2.10.6
rq==0.12.0
six==1.11.0
text-unidecode==1.2
tornado==5.1.1
typed-ast==1.1.0
vine==1.1.4
wrapt==1.10.11
